refactor: Moved routes into a separate directory and bootstrap into the configuration folder:
