+++ matchingService/src/config/bootstrap.ts
@@ -1,8 +1,8 @@
-import QueueService from "./QueueService/QueueService";
-import MatchService from "./services/MatchService";
-import MatchController from "./controllers/MatchController";
-import MatchRequest from "./models/MatchRequest";
-import logger from "./utils/logger";
+import QueueService from "../QueueService/QueueService";
+import MatchService from "../services/MatchService";
+import MatchController from "../controllers/MatchController";
+import MatchRequest from "../models/MatchRequest";
+import logger from "../utils/logger";
 
 export interface IQueueService {
     sendMatchRequest(matchRequest: MatchRequest): Promise<boolean>;
