+++ matchingService/src/routes/matching.ts
@@ -1,7 +1,7 @@
 import express, { Router } from "express";
-import MatchController from "./controllers/MatchController";
+import MatchController from "../controllers/MatchController";
 
-export default function createRouter(controller: MatchController): Router {
+export default function createMatchingRouter(controller: MatchController): Router {
     const router: Router = express.Router();
 
     router.post("/findMatch", (req, res, next) => controller.findMatch(req, res, next));
